requirements.txt
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Pillow==7.0.0
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.3.2
streamlit==0.60.0
tensorflow==1.13.1
tensorflow-estimator==1.13.0